refactor: Re-structure flask app - Create 'app' directory - Create 'blueprints' directory under the app and move Blueprint files to it. - Rename events.py to viewer_event.py - Create 'events' directory under the app and move viewer_event.py to it. - Move static and templates directory under the app
