+++ app/static/css/viewer_style.css
